refactor: Refatora o Controller para incluir ResponseEntity nas rotas para customizar a resposta das requisições do cliente

+++ src/main/java/com/mucida/study/controller/NinjaController.java
@@ -2,6 +2,8 @@ package com.mucida.study.controller;
 
 import com.mucida.study.model.dto.NinjaDTO;
 import com.mucida.study.service.NinjaService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,28 +19,42 @@ public class NinjaController {
     }
 
     @PostMapping()
-    public void addNinja(@RequestBody NinjaDTO ninja){
+    public ResponseEntity<String> addNinja(@RequestBody NinjaDTO ninja){
         ninjaService.addNinja(ninja);
+        return ResponseEntity.status(HttpStatus.CREATED).body(null);
     }
 
     @GetMapping()
-    public List<NinjaDTO> getAllNinjas(){
-        return ninjaService.getAllNinjas();
+    @ResponseStatus(HttpStatus.OK)
+    public ResponseEntity<List<NinjaDTO>> getAllNinjas(){
+        return ResponseEntity.status(HttpStatus.OK).body(ninjaService.getAllNinjas());
     }
 
     @GetMapping("/{id}")
-    public NinjaDTO getNinjaById(@PathVariable Long id){
-        return ninjaService.getNinjaById(id);
+    public ResponseEntity<NinjaDTO> getNinjaById(@PathVariable Long id){
+        if (ninjaService.getNinjaById(id) != null){
+            return ResponseEntity.status(HttpStatus.OK).body(ninjaService.getNinjaById(id));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
 
     @PutMapping("/{id}")
-    public NinjaDTO updateNinja(@PathVariable Long id, @RequestBody NinjaDTO ninja){
-        return ninjaService.updateNinja(id, ninja);
+    public ResponseEntity<NinjaDTO> updateNinja(@PathVariable Long id, @RequestBody NinjaDTO ninja){
+        if (ninjaService.getNinjaById(id) != null){
+            return ResponseEntity.status(HttpStatus.OK).body(ninjaService.updateNinja(id, ninja));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteNinja(@PathVariable Long id){
-        ninjaService.deleteNinja(id);
+    public ResponseEntity<String> deleteNinja(@PathVariable Long id){
+        if (ninjaService.getNinjaById(id) != null){
+            ninjaService.deleteNinja(id);
+            return ResponseEntity.status(HttpStatus.OK).body(null);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
     }
-
 }
